+++ pwsh/dev/functions/processScopeInsightsMgOrSub.ps1
@@ -632,9 +632,9 @@ extensions: [{ name: 'sort' }]
                 $totalCost = 0
 
                 $currency = $htAzureConsumptionSubscriptions.($subscriptionId).Currency
-                $consumedServiceCount = ($consumptionData.consumedService | Sort-Object -Unique | Measure-Object).Count
+                $consumedServiceCount = ($consumptionData.ResourceType | Sort-Object -Unique | Measure-Object).Count
                 $resourceCount = ($consumptionData.ResourceId | Sort-Object -Unique | Measure-Object).Count
-                $subConsumptionDataGrouped = $consumptionData | Group-Object -property ConsumedService, ChargeType, MeterCategory
+                $subConsumptionDataGrouped = $consumptionData | Group-Object -property ResourceType, ChargeType, MeterCategory
 
                 foreach ($consumptionline in $subConsumptionDataGrouped) {
 
@@ -647,7 +647,7 @@ extensions: [{ name: 'sort' }]
                     }
 
                     $null = $arrayConsumptionData.Add([PSCustomObject]@{
-                            ConsumedService              = ($consumptionline.name).split(', ')[0]
+                            ResourceType                 = ($consumptionline.name).split(', ')[0]
                             ConsumedServiceChargeType    = ($consumptionline.name).split(', ')[1]
                             ConsumedServiceCategory      = ($consumptionline.name).split(', ')[2]
                             ConsumedServiceInstanceCount = $consumptionline.Count
@@ -691,7 +691,7 @@ extensions: [{ name: 'sort' }]
                     @"
 <tr>
 <td>$($consumptionLine.ConsumedServiceChargeType)</td>
-<td>$($consumptionLine.ConsumedService)</td>
+<td>$($consumptionLine.ResourceType)</td>
 <td>$($consumptionLine.ConsumedServiceCategory)</td>
 <td>$($consumptionLine.ConsumedServiceInstanceCount)</td>
 <td>$($consumptionLine.ConsumedServiceCost)</td>
@@ -1228,9 +1228,9 @@ extensions: [{ name: 'sort' }]
                     $consumptionDataGroupedByCurrency = $consumptionData | Group-Object -property Currency
                     foreach ($currency in $consumptionDataGroupedByCurrency) {
                         $totalCost = 0
-                        $tenantSummaryConsumptionDataGrouped = $currency.group | Group-Object -property ConsumedService, ChargeType, MeterCategory
+                        $tenantSummaryConsumptionDataGrouped = $currency.group | Group-Object -property ResourceType, ChargeType, MeterCategory
                         $subsCount = ($tenantSummaryConsumptionDataGrouped.group.subscriptionId | Sort-Object -Unique).Count
-                        $consumedServiceCount = ($tenantSummaryConsumptionDataGrouped.group.consumedService | Sort-Object -Unique).Count
+                        $consumedServiceCount = ($tenantSummaryConsumptionDataGrouped.group.ResourceType | Sort-Object -Unique).Count
                         $resourceCount = ($tenantSummaryConsumptionDataGrouped.group.ResourceId | Sort-Object -Unique).Count
                         foreach ($consumptionline in $tenantSummaryConsumptionDataGrouped) {
 
@@ -1243,7 +1243,7 @@ extensions: [{ name: 'sort' }]
                             }
 
                             $null = $arrayConsumptionData.Add([PSCustomObject]@{
-                                    ConsumedService              = ($consumptionline.name).split(', ')[0]
+                                    ResourceType                 = ($consumptionline.name).split(', ')[0]
                                     ConsumedServiceChargeType    = ($consumptionline.name).split(', ')[1]
                                     ConsumedServiceCategory      = ($consumptionline.name).split(', ')[2]
                                     ConsumedServiceInstanceCount = $consumptionline.Count
@@ -1291,7 +1291,7 @@ extensions: [{ name: 'sort' }]
                         @"
 <tr>
 <td>$($consumptionLine.ConsumedServiceChargeType)</td>
-<td>$($consumptionLine.ConsumedService)</td>
+<td>$($consumptionLine.ResourceType)</td>
 <td>$($consumptionLine.ConsumedServiceCategory)</td>
 <td>$($consumptionLine.ConsumedServiceInstanceCount)</td>
 <td>$($consumptionLine.ConsumedServiceCost)</td>
@@ -3237,6 +3237,94 @@ extensions: [{ name: 'sort' }]
 <tr><td class="detailstd">
 '@)
     #endregion ScopeInsightsBlueprintsScoped
+
+    if ($mgOrSub -eq 'sub') {
+        #region ScopeInsightsClassicAdministrators
+        if ($htClassicAdministrators.($subscriptionId).ClassicAdministrators.Count -gt 0) {
+            $tfCount = $htClassicAdministrators.($subscriptionId).ClassicAdministrators.Count
+            $htmlTableId = "ScopeInsights_ClassicAdministrators_$($subscriptionId -replace '\(','_' -replace '\)','_' -replace '-','_' -replace '\.','_')"
+            $randomFunctionName = "func_$htmlTableId"
+            [void]$htmlScopeInsights.AppendLine(@"
+<button onclick="loadtf$("func_$htmlTableId")()" type="button" class="collapsible"><p><i class="fa fa-check-circle blue" aria-hidden="true"></i> $tfCount Classic Administrators</p></button>
+<div class="content $SIDivContentClass">
+&nbsp;&nbsp;<i class="fa fa-table" aria-hidden="true"></i> Download CSV <a class="externallink" href="#" onclick="download_table_as_csv_semicolon('$htmlTableId');">semicolon</a> | <a class="externallink" href="#" onclick="download_table_as_csv_comma('$htmlTableId');">comma</a>
+<table id="$htmlTableId" class="$cssClass">
+<thead>
+<tr>
+<th>Role</th>
+<th>Identity</th>
+</tr>
+</thead>
+<tbody>
+"@)
+            $htmlScopeInsightsClassicAdministrators = $null
+            $htmlScopeInsightsClassicAdministrators = foreach ($classicAdministrator in $htClassicAdministrators.($subscriptionId).ClassicAdministrators | Sort-Object -Property Role, Identity) {
+                @"
+<tr>
+<td>$($classicAdministrator.Role)</td>
+<td>$($classicAdministrator.Identity)</td>
+</tr>
+"@
+            }
+            [void]$htmlScopeInsights.AppendLine($htmlScopeInsightsClassicAdministrators)
+            [void]$htmlScopeInsights.AppendLine(@"
+                </tbody>
+            </table>
+        </div>
+        <script>
+            function loadtf$("func_$htmlTableId")() { if (window.helpertfConfig4$htmlTableId !== 1) {
+                window.helpertfConfig4$htmlTableId =1;
+                var tfConfig4$htmlTableId = {
+                base_path: 'https://www.azadvertizer.net/azgovvizv4/tablefilter/', rows_counter: true,
+"@)
+            if ($tfCount -gt 10) {
+                $spectrum = "10, $tfCount"
+                if ($tfCount -gt 50) {
+                    $spectrum = "10, 25, 50, $tfCount"
+                }
+                if ($tfCount -gt 100) {
+                    $spectrum = "10, 30, 50, 100, $tfCount"
+                }
+                if ($tfCount -gt 500) {
+                    $spectrum = "10, 30, 50, 100, 250, $tfCount"
+                }
+                if ($tfCount -gt 1000) {
+                    $spectrum = "10, 30, 50, 100, 250, 500, 750, $tfCount"
+                }
+                if ($tfCount -gt 2000) {
+                    $spectrum = "10, 30, 50, 100, 250, 500, 750, 1000, 1500, $tfCount"
+                }
+                if ($tfCount -gt 3000) {
+                    $spectrum = "10, 30, 50, 100, 250, 500, 750, 1000, 1500, 3000, $tfCount"
+                }
+                [void]$htmlScopeInsights.AppendLine(@"
+paging: {results_per_page: ['Records: ', [$spectrum]]},/*state: {types: ['local_storage'], filters: true, page_number: true, page_length: true, sort: true},*/
+"@)
+            }
+            [void]$htmlScopeInsights.AppendLine(@"
+                btn_reset: true, highlight_keywords: true, alternate_rows: true, auto_filter: { delay: 1100 }, no_results_message: true,
+                col_types: [
+                    'caseinsensitivestring',
+                    'caseinsensitivestring'
+                ],
+                extensions: [{ name: 'sort' }]
+            };
+            var tf = new TableFilter('$htmlTableId', tfConfig4$htmlTableId);
+            tf.init();}}
+        </script>
+"@)
+        }
+        else {
+            [void]$htmlScopeInsights.AppendLine(@"
+                    <p><i class="fa fa-ban" aria-hidden="true"></i> No Classic Administrators</p>
+"@)
+        }
+        [void]$htmlScopeInsights.AppendLine(@'
+</td></tr>
+<tr><td class="detailstd">
+'@)
+        #endregion ScopeInsightsClassicAdministrators
+    }
 
     #RoleAssignments
     #region ScopeInsightsRoleAssignments